spaced-repetition/backend/requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
djongo==1.2.31
dnspython==2.7.0
pymongo==4.11.3
sqlparse==0.5.3
